refactor: added the feat packages to the test package
